Start of basic physics velocity & acceleration implementation
Added the update method to mjObject and the skeleton of responding to
effects, though the effect object itself does not exist. Probably the
next step.
Also added part of the skeleton for the collision detection and object
update cycle .
Hopefully soon I get to have the same possibilities as in the Java
version.
The controls will likely be problematic to implement thanks to the
"audacity" of Android sdk/api developers since it involves using an
undocumented function