squished bugs in lookAt and MultiplyMM, did some work on mjBoundingStructures
FINALLY squished ALL the lookAt matrix bugs. apparently.
Corrected the matrix multiplication to actually do POSTmultiplication
and NOT PREmultiplication
New mjBoundingStructures should by default be a Sphere with a radius 0
(a particle). This will be enforced eventually. Also at some point I
will check whether an actual mjBoundingStructure for particle is
necessary or we can run with the infinitesimal spheres without problems