AABBvsAABB collisions detected correctly, response not quite
collisions between AABBs are detected right but the response has
something weird. (one object seems to drag the other around - one
becomes subordinated to the other. Which is not correct)

I suspect the adjustment and the object response to physical effects
might have "details"

Getting to this state required some adjustments to mjObject so that it
can handle mjAABBs on its own.