Created desktop test environment; cleanup
created a test environment to be able to run the program in the Desktop.
It does not show things yet, but that's probably just some GL
initialisation parameter missing, since part of the setup resides on the
Java side in Android.

It also probably only works in GNU/Linux with libpng16 hand-compiled and
installed.
Based on the printf messages, other than not showing any graphics, the
rest of the engine seems to be running.

Also it is amazing it compiled at all before, it was missing tons of
includes and had a couple of typos.