¿Quién es Sombra?... started with mjGraphicText; added mjEngine lib for pc-linux-x86_64
mjGraphicText will be used to render phrases and possibly even
paragraphs in the future. It creates a vector containing
mjGraphicCharObjects that can be individually coloured and displaced.
Right now it doesn't size the characters properly, but that will be
investigated tomorrow and is likely to not be simple.
Updating the text properly will require some brain too ^^~

Decided to start adding libmjEngine.so as a precompiled library too,
after some path fiasco caused by a super-old qmake bug that seems to
either not be fixable or not important enough for the Qt people to fix.