added non-immediate mode to mjGraphicText and ACTION_UP to Takkatakka.java (bugfix)
Added non-immediate mode to mjGraphicText.
By default, mjGraphicText will immediately try to update its internal
mjGraphicChar:s, however if this is called from a thread without a GL
context, it will fail and leave empty textures behind, which will be
shown (invisibly -_-) whenever that specific char is called into action.
Which means it screws everything for everyone :D

To avoid this, one can use the non-immediate mode instead, then remember
to call Update() on the mjGraphicText from a thread with a GL context
(the universe's Update() method always has a GL context, so this is the
ideal place).

--

Takkatakka wasn't sending ACTION_UP events due to an oversight, which is
kind of a big deal when counting fingers pressing the screen.. :3