Made the object lists in mjSceneGraph private; added/optimised Add and Remove methods
the object lists in mjSceneGraph are now private and adding and removing
objects from them has been optimised. Now the object lookup is not
necessary in most cases because the object knows in which location it
is.
The AddGroup/RemoveGroup options are for the time being commented,
because they are not in use and the optimisation has not been written
for them yet.