The sceneGraph optimisation wasn't so clever after all
The logic behind the sceneGraph optimisation was significantly flawed.
For the time being I don't see how it can be brought back so I'll
probably remove it by the next commit.

However, the game developers can make an educated guess and either start
to look for elements to remove from the first or the last one, depending
on where they think the element they need to remove from the scenegraph
will be.