Windows-required changes
Updates to .pro file to properly find the files in Windows; also
eliminated an ambiguity issue between Qt and TinyXML2 despite Qt not
being at all in the namespace.
Added a function to change "/" to windows directory separators ("\")
when running under Windows. mjEngine requires you to _always_ use "/"
for simplicity.