Changes to fully use mjInterfaceBuilder and pokeable objects
A new mjObject subclass, mjUIObject, has been created. This one defines
a "pokeable" surface within the UI coordinates space: [-1,1] for the
smaller dimension, and beyond -1,1 for the larger dimension. i.e. the
smaller dimension for a device in landscape mode is the VERTICAL
dimension, Y; in portrait mode the HORIZONTAL dimension, X, is the
smaller one.

When poked, the object emits a message that by default is
<objectID>.poked, but can be customised via XML attributes.

The setting of XML attributes will probably have to be re-examined in
the future.